fix: dynamic build error
[#145] correção rota /empreendimentos/cadastrar

https://jira.arinterface.com.br/work_packages/145
